[restructure] files
