+++ src/components/Loading/LoadingSpinner.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import './styles.scss';
+
+function Spinner() {
+	return (
+		<div className="loader">Loading...</div>
+	)
+}
+
+export default Spinner
